+++ autogluon/model_zoo/model_store.py
@@ -6,7 +6,7 @@ __all__ = ['get_model_file', 'purge']
 import os
 import zipfile
 
-from ..utils import download, check_sha1
+from autogluon_core.utils import download, check_sha1
 
 
 _model_sha1 = {name: checksum for checksum, name in [
