+++ autogluon/task/image_classification/classifier.py
@@ -14,7 +14,7 @@ from .nets import get_network
 from .utils import *
 from ..base.base_predictor import BasePredictor
 from autogluon_core.core import AutoGluonObject
-from ...utils import save, load, tqdm, collect_params, update_params
+from autogluon_core.utils import save, load, tqdm, collect_params, update_params
 
 __all__ = ['Classifier']
