+++ autogluon/task/object_detection/detector.py
@@ -13,7 +13,7 @@ from mxnet import gluon
 from .utils import rcnn_split_and_load, get_network
 from ..base.base_predictor import BasePredictor
 from autogluon_core.core import AutoGluonObject
-from ...utils import save, load, collect_params, update_params
+from autogluon_core.utils import save, load, collect_params, update_params
 
 __all__ = ['Detector']
