+++ autogluon/utils/tabular/metrics/__init__.py
@@ -9,7 +9,7 @@ from sklearn.utils.multiclass import type_of_target
 from . import classification_metrics, softclass_metrics
 from .util import sanitize_array
 from ..ml.constants import PROBLEM_TYPES, PROBLEM_TYPES_REGRESSION, PROBLEM_TYPES_CLASSIFICATION
-from ...miscs import warning_filter
+from autogluon_core.utils.miscs import warning_filter
 
 
 class Scorer(object, metaclass=ABCMeta):
