+++ autogluon/utils/tabular/ml/models/lgb/lgb_model.py
@@ -19,7 +19,7 @@ from .lgb_utils import construct_dataset
 from ..abstract.abstract_model import AbstractModel, fixedvals_from_searchspaces
 from ...constants import BINARY, MULTICLASS, REGRESSION
 from ....utils.savers import save_pkl
-from .....try_import import try_import_lightgbm
+from autogluon_core.utils.try_import import try_import_lightgbm
 from autogluon_core.core import Int, Space
 
 warnings.filterwarnings("ignore", category=UserWarning, message="Starting from version")  # lightGBM brew libomp warning
