+++ autogluon_core/core/decorator.py
@@ -9,8 +9,8 @@ import ConfigSpace as CS
 
 from .space import *
 from .space import _add_hp, _add_cs, _rm_hp, _strip_config_space
-from autogluon.utils import EasyDict as ezdict
-from autogluon.utils.deprecate import make_deprecate
+from ..utils import EasyDict as ezdict
+from ..utils.deprecate import make_deprecate
 
 __all__ = ['args', 'obj', 'func', 'sample_config',
            'autogluon_register_args', 'autogluon_object', 'autogluon_function',
