+++ autogluon_core/scheduler/remote/remote_manager.py
@@ -4,7 +4,7 @@ import logging
 import multiprocessing as mp
 
 from .remote import Remote
-from autogluon.utils import warning_filter
+from ...utils import warning_filter
 
 __all__ = ['RemoteManager']
